DDF-04357 Split into modules
